requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
asyncio>=3.4.3
websockets>=10.0
requests>=2.25.0

# Trading and financial data
python-binance>=1.0.16
ccxt>=2.5.0
ta>=0.7.0
TA-Lib>=0.4.21

# Machine Learning and GPU acceleration
torch>=1.12.0
torchvision>=0.13.0
tensorflow>=2.9.0
scikit-learn>=1.0.0
xgboost>=1.6.0
lightgbm>=3.3.0
cupy-cuda12x>=10.0.0  # GPU-accelerated NumPy for CUDA 12.x
numba>=0.56.0  # JIT compilation

# Data processing and analysis
scipy>=1.7.0
statsmodels>=0.13.0
plotly>=5.0.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Configuration and utilities
PyYAML>=6.0
python-dotenv>=0.19.0
loguru>=0.6.0
rich>=12.0.0
typer>=0.6.0

# Performance monitoring
psutil>=5.8.0
memory-profiler>=0.60.0

# Database (optional)
redis>=4.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.19.0

# Streamlit for GUI dashboard
streamlit>=1.35.0 